war reinforcement text display fix + "grace period" for initially taken spawns (period duration controlled by modoption)

+++ src/persistence/saveable_objs/TurfZone.cs
@@ -79,7 +79,7 @@ namespace GTA.GangAndTurfMod
                         Function.Call(Hash.SET_BLIP_SECONDARY_COLOUR, myBlip, 255, 0f, 0f);
                     }
 
-                    myBlip.Scale = 1.0f + 0.65f / ((ModOptions.instance.maxTurfValue + 1) / (value + 1));
+                    myBlip.Scale = 1.0f + 0.65f / ((ModOptions.instance.maxTurfValue + 1) / (float) (value + 1));
                 }
 
                 Function.Call(Hash.BEGIN_TEXT_COMMAND_SET_BLIP_NAME, "STRING");
